Implement input error hadnling part 1

+++ src/main/java/parkinglot/controller/DeleteController.java
@@ -10,6 +10,8 @@ import parkinglot.service.ParkingPlaceService;
 import parkinglot.service.ParkingService;
 import parkinglot.service.ParkingZoneService;
 
+import static parkinglot.config.Messages.INVALID_ID;
+
 @Controller
 public class DeleteController {
     private final ParkingService parkingService;
@@ -38,21 +40,38 @@ public class DeleteController {
         return "select-parking-delete";
     }
     @PostMapping("/select-parking-delete")
-    public String deleteParking (Model model,Long parkingId){
-        boolean parkingPresent = parkingService.isParkingPresent(parkingId);
-        if (!parkingPresent){
-            String text = "Parking with Id " + parkingId + " doesn't exist!";
-            model.addAttribute("infoText",text);
-        } else {
-            parkingService.deleteById(parkingId);
+    public String deleteParking (Model model,String StringId) {
+        boolean catchException = false;
+        try {
+            Long parkingId = Long.valueOf(StringId);
+        } catch (NumberFormatException nfe) {
+            catchException = true;
         }
-        String allParkings = parkingService.getAllParkings();
-        if (allParkings.isEmpty()){
-            allParkings="Don't have parkings, if you want to add you can go in parking operation and add Parking";
-        }
-        model.addAttribute("parkingList", allParkings);
+        if (!catchException) {
+            Long parkingId = Long.valueOf(StringId);
+            boolean parkingPresent = parkingService.isParkingPresent(parkingId);
+            if (!parkingPresent) {
+                model.addAttribute("mistakeForId",INVALID_ID);
+            } else {
+                parkingService.deleteById(parkingId);
+            }
+            String allParkings = parkingService.getAllParkings();
+            if (allParkings.isEmpty()) {
+                allParkings = "Don't have parkings, if you want to add you can go in parking operation and add Parking";
+            }
+            model.addAttribute("parkingList", allParkings);
 
-        return "select-parking-delete";
+            return "select-parking-delete";
+        }else {
+            String allParkings = parkingService.getAllParkings();
+            if (allParkings.isEmpty()) {
+                allParkings = "Don't have parkings, if you want to add you can go in parking operation and add Parking";
+            }
+            model.addAttribute("parkingList", allParkings);
+            model.addAttribute("mistakeForId", INVALID_ID);
+
+            return "select-parking-delete";
+        }
     }
 
     @GetMapping("/select-zone-delete")
@@ -67,22 +86,40 @@ public class DeleteController {
         return "select-zone-delete";
     }
     @PostMapping("/select-zone-delete")
-    public String deleteParkingZone (Model model,Long zoneId) {
-        boolean zonePresent = parkingZoneService.isZonePersent(zoneId);
-        if (!zonePresent) {
-            String text = "Zone with Id:  " + zoneId + " doesn't exist!";
-            model.addAttribute("infoText", text);
+    public String deleteParkingZone (Model model,String StringId) {
+        boolean catchException = false;
+        try {
+            Long zoneId = Long.valueOf(StringId);
+        } catch (NumberFormatException nfe) {
+            catchException = true;
+        }
+        if (!catchException) {
+            Long zoneId = Long.valueOf(StringId);
+            boolean zonePresent = parkingZoneService.isZonePersent(zoneId);
+            if (!zonePresent) {
+                model.addAttribute("mistakeForId",INVALID_ID);
+            } else {
+                parkingZoneService.deleteZone(zoneId);
+            }
+            String allZone = parkingZoneService.getAllZones();
+
+            if (allZone.isEmpty()) {
+                allZone = "Don't have zones, if you want to add you can go in Parking Zone operation and add Zone";
+
+            }
+            model.addAttribute("parkingZones", allZone);
+            return "select-zone-delete";
         } else {
-            parkingZoneService.deleteZone(zoneId);
-        }
-        String allZone = parkingZoneService.getAllZones();
+            String allZone = parkingZoneService.getAllZones();
 
-        if (allZone.isEmpty()) {
-            allZone = "Don't have zones, if you want to add you can go in Parking Zone operation and add Zone";
+            if (allZone.isEmpty()) {
+                allZone = "Don't have zones, if you want to add you can go in Parking Zone operation and add Zone";
 
+            }
+            model.addAttribute("mistakeForId", INVALID_ID);
+            model.addAttribute("parkingZones", allZone);
+            return "select-zone-delete";
         }
-        model.addAttribute("parkingZones", allZone);
-        return "select-zone-delete";
     }
 
     @GetMapping("/select-place-delete")
@@ -97,21 +134,37 @@ public class DeleteController {
         return "select-place-delete";
     }
     @PostMapping("/select-place-delete")
-    public String deleteParkingPlace(Model model,Long placeId){
-        boolean placePresent = parkingPlaceService.isPlacePresent(placeId);
-        if (!placePresent) {
-            String text = "Parking Place with Id:  " + placeId + " doesn't exist!";
-            model.addAttribute("infoText", text);
-        } else {
-            parkingPlaceService.deletePlace(placeId);
+    public String deleteParkingPlace(Model model,String StringId) {
+        boolean catchException = false;
+        try {
+            Long placeId = Long.valueOf(StringId);
+        } catch (NumberFormatException nfe) {
+            catchException = true;
         }
-        String allPLaces = parkingPlaceService.getAllParkingPLaces();
-        if (allPLaces.isEmpty()){
-            allPLaces="Don't have places, if you want to add you can go in Parking Place operation and add Place";
-        }
-        model.addAttribute("parkingPlaces", allPLaces);
+        if (!catchException) {
+            Long placeId = Long.valueOf(StringId);
+            boolean placePresent = parkingPlaceService.isPlacePresent(placeId);
+            if (!placePresent) {
+                model.addAttribute("mistakeForId", INVALID_ID);
+            } else {
+                parkingPlaceService.deletePlace(placeId);
+            }
+            String allPLaces = parkingPlaceService.getAllParkingPLaces();
+            if (allPLaces.isEmpty()) {
+                allPLaces = "Don't have places, if you want to add you can go in Parking Place operation and add Place";
+            }
+            model.addAttribute("parkingPlaces", allPLaces);
 
-        return "select-place-delete";
+            return "select-place-delete";
+        }else {
+            String allPLaces = parkingPlaceService.getAllParkingPLaces();
+            if (allPLaces.isEmpty()) {
+                allPLaces = "Don't have places, if you want to add you can go in Parking Place operation and add Place";
+            }
+            model.addAttribute("parkingPlaces", allPLaces);
+            model.addAttribute("mistakeForId", INVALID_ID);
+            return "select-place-delete";
+        }
     }
     @GetMapping("/select-car-delete")
     public String exportCars(Model model) {
@@ -125,11 +178,18 @@ public class DeleteController {
     }
 
     @PostMapping("/select-car-delete")
-    public String deleteCaR (Model model,Long carId){
-            boolean carPresent = carService.isCarPresent(carId);
+    public String deleteCaR (Model model,String StringId){
+        boolean catchException = false;
+        try {
+            Long carId = Long.valueOf(StringId);
+        } catch (NumberFormatException nfe) {
+            catchException = true;
+        }
+        if (!catchException) {
+            Long carId = Long.valueOf(StringId);
+        boolean carPresent = carService.isCarPresent(carId);
             if (!carPresent) {
-                String text = "Car with Id:  " + carId + " doesn't exist!";
-                model.addAttribute("infoText", text);
+                model.addAttribute("mistakeForId", INVALID_ID);
             } else {
                 carService.deleteCar(carId);
             }
@@ -139,5 +199,14 @@ public class DeleteController {
         }
         model.addAttribute("Cars", allCars);
         return "select-car-delete";
+        }else {
+            String allCars = carService.getAllCars();
+            if (allCars.isEmpty()){
+                allCars="Don't have cars, if you want to add you can go in Car Operation and add Car";
+            }
+            model.addAttribute("Cars", allCars);
+            model.addAttribute("mistakeForId", INVALID_ID);
+            return "select-car-delete";
+        }
     }
 }
